remove public place in lib
